+++ test/src/test/java/com/pwinckles/jdbcgen/test/UuidIdEntityTest.java
@@ -12,7 +12,7 @@ import org.apache.commons.lang3.tuple.Pair;
 
 public class UuidIdEntityTest
         extends TestBase<
-                UuidIdEntity, UUID, UuidIdEntityDb.Patch, UuidIdEntityDb.Column, UuidIdEntityDb.FilterBuilder> {
+                UuidIdEntity, UUID, UuidIdEntityDb.Patch, UuidIdEntityDb.FilterBuilder, UuidIdEntityDb.SortBuilder> {
 
     // Need to do this so the uuids are ordered
     private final Iterator<UUID> uuids = Stream.generate(UUID::randomUUID)